Final submission/Command line program/requirements.txt
torch==0.4.0
torchvision==0.2.1
numpy==1.12.1
Pillow==5.2.0